refactor: config.ts を分割し各パッケージに移動
- McAuthMode / parseMcAuthMode → packages/minecraft/constants.ts
- HEARTBEAT_CONFIG_RELATIVE_PATH → packages/scheduling/heartbeat-helpers.ts
- loadConfig / AppConfig / TtsConfig / MinecraftConfig → apps/discord/src/config.ts
- shared/config.ts は APP_ROOT のみ残す

Refs #277
