+++ apps/discord/src/bootstrap.test.ts
@@ -1,8 +1,7 @@
 import { describe, expect, mock, test } from "bun:test";
 
-import type { AppConfig } from "@vicissitude/shared/config";
-
 import { createStoreLayer, createMetrics } from "./bootstrap.ts";
+import type { AppConfig } from "./config.ts";
 
 function createTestConfig(overrides?: Partial<AppConfig>): AppConfig {
 	return {
