+++ apps/discord/src/bootstrap.ts
@@ -26,12 +26,8 @@ import { OllamaEmbeddingAdapter } from "@vicissitude/ollama";
 import { OPENCODE_ALL_TOOLS_DISABLED } from "@vicissitude/opencode/constants";
 import { OpencodeSessionAdapter } from "@vicissitude/opencode/session-adapter";
 import { ConsolidationScheduler } from "@vicissitude/scheduling/consolidation-scheduler";
+import { HEARTBEAT_CONFIG_RELATIVE_PATH } from "@vicissitude/scheduling/heartbeat-helpers";
 import { HeartbeatScheduler } from "@vicissitude/scheduling/heartbeat-scheduler";
-import {
-	type AppConfig,
-	HEARTBEAT_CONFIG_RELATIVE_PATH,
-	loadConfig,
-} from "@vicissitude/shared/config";
 import type {
 	AiAgent,
 	ContextBuilderPort,
@@ -45,6 +41,7 @@ import { incrementEmoji } from "@vicissitude/store/queries";
 import { createAivisSpeechSynthesizer, createEmotionToTtsStyleMapper } from "@vicissitude/tts";
 import { spawn, type Subprocess } from "bun";
 
+import { type AppConfig, loadConfig } from "./config.ts";
 import { ChannelConfigLoader, type ChannelConfigData } from "./gateway/channel-config-loader.ts";
 import { DiscordGateway } from "./gateway/discord.ts";
