+++ packages/scheduling/src/heartbeat-scheduler.ts
@@ -2,12 +2,11 @@ import { resolve } from "path";
 
 import { HeartbeatService } from "@vicissitude/application/heartbeat-service";
 import { METRIC } from "@vicissitude/observability/metrics";
-import { HEARTBEAT_CONFIG_RELATIVE_PATH } from "@vicissitude/shared/config";
 import { delayResolve, withTimeout } from "@vicissitude/shared/functions";
 import type { AiAgent, HeartbeatConfig, Logger, MetricsCollector } from "@vicissitude/shared/types";
 
 import { JsonHeartbeatConfigRepository } from "./heartbeat-config.ts";
-import { evaluateDueReminders } from "./heartbeat-helpers.ts";
+import { HEARTBEAT_CONFIG_RELATIVE_PATH, evaluateDueReminders } from "./heartbeat-helpers.ts";
 
 // ─── HeartbeatScheduler ─────────────────────────────────────────
